+++ tests/Hasher/HasherTest.php
@@ -18,6 +18,7 @@ namespace Esi\ConsistentHash\Tests\Hasher;
 use Esi\ConsistentHash\ConsistentHash;
 use Esi\ConsistentHash\Hasher\Crc32Hasher;
 use Esi\ConsistentHash\Hasher\Md5Hasher;
+use Esi\ConsistentHash\Hasher\Murmur3Hasher;
 use Esi\ConsistentHash\Hasher\Xxh32Hasher;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\UsesClass;
@@ -28,6 +29,7 @@ use PHPUnit\Framework\TestCase;
  */
 #[CoversClass(Crc32Hasher::class)]
 #[CoversClass(Md5Hasher::class)]
+#[CoversClass(Murmur3Hasher::class)]
 #[CoversClass(Xxh32Hasher::class)]
 #[UsesClass(ConsistentHash::class)]
 class HasherTest extends TestCase
@@ -80,6 +82,35 @@ class HasherTest extends TestCase
         }
 
         $hashSpace = new ConsistentHash(new Md5Hasher());
+        $hashSpace->addTargets($targets);
+
+        foreach (range(1, 10) as $i) {
+            self::assertTrue(
+                \in_array($hashSpace->lookup(\sprintf('r%s', $i)), $targets, true),
+                'target must be in list of targets',
+            );
+        }
+    }
+
+    public function testMurmur3Hash(): void
+    {
+        $hasher  = new Murmur3Hasher();
+        $result1 = $hasher->hash('test');
+        $result2 = $hasher->hash('test');
+        $result3 = $hasher->hash('different');
+
+        self::assertEquals($result1, $result2);
+        self::assertNotEquals($result1, $result3); // fragile but worthwhile
+    }
+
+    public function testMurmur3HashSpaceLookupsAreValidTargets(): void
+    {
+        $targets = [];
+        foreach (range(1, 10) as $i) {
+            $targets[] = \sprintf('target%s', $i);
+        }
+
+        $hashSpace = new ConsistentHash(new Murmur3Hasher());
         $hashSpace->addTargets($targets);
 
         foreach (range(1, 10) as $i) {