+++ tests/Bench/LookupBench.php
@@ -16,7 +16,12 @@ declare(strict_types=1);
 namespace Esi\ConsistentHash\Tests\Bench;
 
 use Esi\ConsistentHash\ConsistentHash;
+use Esi\ConsistentHash\Hasher\HasherInterface;
+use Esi\ConsistentHash\Hasher\Crc32Hasher;
+use Esi\ConsistentHash\Hasher\Fnv1AHasher;
 use Esi\ConsistentHash\Hasher\Md5Hasher;
+use Esi\ConsistentHash\Hasher\Murmur3Hasher;
+use Esi\ConsistentHash\Hasher\Xxh32Hasher;
 use Generator;
 use PhpBench\Attributes\ParamProviders;
 
@@ -33,12 +38,6 @@ final class LookupBench
 
     public function __construct()
     {
-        $this->hasher = new ConsistentHash(new Md5Hasher(), 64);
-
-        foreach (range(1, 10) as $i) {
-            $this->hasher->addTarget('target_' . $i, 1);
-        }
-
         // Generate random lookup keys outside the measurement function.
         foreach (range(1, 100_000) as $ignored) {
             $this->randomKeys[] = bin2hex(random_bytes(12));
@@ -46,19 +45,29 @@ final class LookupBench
     }
 
     /**
-     * @param array{count: int} $params
+     * @param array{algo: class-string<HasherInterface>, count: int} $params
      */
-    #[ParamProviders(['provideLookupCount', 'provideHasher'])]
+    #[ParamProviders(['provideLookupCount', 'provideHashers'])]
     public function benchLookup(array $params): void
     {
+        $this->hasher = new ConsistentHash(new $params['algo'](), 64);
+
+        foreach (range(1, 10) as $i) {
+            $this->hasher->addTarget('target_' . $i, 1);
+        }
+
         foreach (range(0, $params['count'] - 1) as $i) {
             $this->hasher->lookup($this->randomKeys[$i]);
         }
     }
 
-    public function provideHasher(): Generator
+    public function provideHashers(): Generator
     {
-        yield 'MD5' => ['algo' => 'MD5'];
+        yield 'CRC32' => ['algo' => Crc32Hasher::class];
+        yield 'FNV1A' => ['algo' => Fnv1AHasher::class];
+        yield 'MD5' => ['algo' => Md5Hasher::class];
+        yield 'MURMUR3' => ['algo' => Murmur3Hasher::class];
+        yield 'XXH32' => ['algo' => Xxh32Hasher::class];
     }
 
     public function provideLookupCount(): Generator