+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [Unreleased]
+## [1.1.0] - 2024-12-10
 
 ### Added
 
@@ -86,4 +86,5 @@ This library is based on a fork of [flexihash/flexihash](https://github.com/pda/
 
 
 [unreleased]: https://github.com/ericsizemore/consistent-hash/tree/main
+[1.0.0]: https://github.com/ericsizemore/consistent-hash/releases/tag/1.1.0
 [1.0.0]: https://github.com/ericsizemore/consistent-hash/releases/tag/1.0.0